# swagger.yaml
...
swagger: "2.0"
info:
  description: "Politico-App"
  version: 1.0.0
  title: Politico-App
  termsOfService: http://swagger.io/terms/
  #My Contact Details
  contact:
    email: pascalulor@yahoo.com
basePath: /api/v1
...
schemes:
- https
- http
paths:
  /parties:
    post:
      tags:
      - parties
      summary: Add a new political party to the platform
      operationId: add-party
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      produces:
      - application/json
      
      parameters:
      - in: body
        name: name
        description: The name of the party that to be create
        required: true
        # schema:
        #   $ref: '#/definitions/Pet'
      - in: body
        name: email
        description: The email of the party that to be create
        required: true
        # schema:
        #   $ref: '#/definitions/Pet'
      - in: body
        name: hqAddress
        description: The contact address of the party that to be create
        required: true
        # schema:
        #   $ref: '#/definitions/Pet'
      - in: body
        name: phonenumber
        description: The phonenumber of the party that to be create
        required: true
        # schema:
        #   $ref: '#/definitions/Pet'
      - in: body
        name: logoUrl
        description: The logo image of the party that to be create
        required: false
        # schema:
        #   $ref: '#/definitions/Pet'
      - in: body
        name: about
        description: brief text about the party that to be create
        required: false
        # schema:
        #   $ref: '#/definitions/Pet'
      responses:
        201:
          description: party created successfully
          examples:
            application/json:
              error: false
              message:
              parties:
                id: 2
                name: DORO MEGA PARTY
                userid: 1
                hqaddress: 32 epic road
                about: wsfsvfffefefewee fefefe
                email: trump@gmail.com
                phonenumber: 8169568499
                logourl: uhhk.jpg
                createdat: 2019-02-07T09:54:29.195Z
                updatedat: 2019-02-07T09:54:29.195Z
        400:
          description: Invalid input
          examples:
            application/json:
              error: true
              message: 
                name: name can only be alphabetical
                hqAddress: Invalid hqAddress
                email: Invalid email
                phonenumber: Invalid phonenumber
                about: about field must be between 20 to 1000 characters
                logoUrl: Invalid logoUrl
        500:
          description: duplicate phonenumber input
          examples:
            application/json:
              error: true
              message: 
                phonenumber: User with phonenumber already exist
      # responses:
      #   400:
      #     description: Invalid input
      security:
      - parties_auth:
        - write:token
        - read:token
    patch:
      tags:
      - parties
      summary: Update an existing party's name
      operationId: update party name
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      produces:
      - application/json
      parameters:
      - in: body
        name: name
        description: Update the name of an existing party
        required: true
        schema:
          $ref: '#/definitions/Pet'
      responses:
        400:
          description: Invalid ID supplied
        404:
          description: Pet not found
        405:
          description: Validation exception
      security:
      - petstore_auth:
        - write:pets
        - read:pets
...
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: http://petstore.swagger.io/oauth/dialog
    flow: implicit
    scopes:
      write:pets: modify pets in your account
      read:pets: read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Category
...
  Tag:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Tag
  Pet:
    type: object
    required:
    - name
    - photoUrls
    properties:
      id:
        type: integer
        format: int64
      category:
        $ref: '#/definitions/Category'
      name:
        type: string
        example: doggie
      photoUrls:
        type: array
        xml:
          name: photoUrl
          wrapped: true
        items:
          type: string
      tags:
        type: array
        xml:
          name: tag
          wrapped: true
        items:
          $ref: '#/definitions/Tag'
      status:
        type: string
        description: pet status in the store
        enum:
        - available
        - pending
        - sold
    xml:
      name: Pet
...